requirements.txt
Flask==3.1.1
psycopg2-binary==2.9.10
pydantic==2.11.7
requests==2.32.4

